% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcem.R
\name{amcem}
\alias{amcem}
\title{Implementing average-simulation-maximization(asm) algorithm by coordinate-wise conditional maximization
Average of parameters for use in the multiple imputations in each iteration.}
\usage{
amcem(
  dat = dat,
  miscode = c(-9),
  method = "sp",
  nem = 10,
  nimpute = 10,
  stepsize = 0.3,
  nstep = 20,
  disp = c(1, 1)
)
}
\arguments{
\item{dat}{a data matrix with possible missing values}

\item{miscode}{a set of values used for denoting missing values}

\item{method}{specific method of estimation, can be
method='sp' for semiparametric likelihood approach (default)
method='pw' for pairwise likelihood approach
method='pm' for permutation likelihood approach}

\item{nem}{the number of EM steps}

\item{nimpute}{the number of Monte Carlo copies for each missing value}

\item{stepsize}{controls the step size of each M-step update.}

\item{nstep}{controls the number of steps before the change of step size}

\item{disp}{controls the layout of the parameter history plot: a two-dimensional
vector of integers in par, mfrow=disp=c(1,1)}
}
\value{
\enumerate{
\item parameter estimate for theta
2. one copy of the imputed data
3. parameter estimates for theta in successful iterations.
}
}
\description{
Implementing average-simulation-maximization(asm) algorithm by coordinate-wise conditional maximization
Average of parameters for use in the multiple imputations in each iteration.
}
\details{
This function maximizes the conditional likelihood coordinate-wise
by the Monte Carlo EM algorithm. In M-step, all the OR parameters are
estimated by the semiparametric OR model using the semiparametric likelihood
approach based on the current imputed data. In MC-step, baseline is first
obtained. Imputing missing values are done coordinate-wise by
the sampling approach
}
\examples{
\dontrun{
 amcem(dat=bone,miss=c(-9),nimpute=100)
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.
}
