% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Compute baseline function estimator given the OR parameters}
\usage{
baseline(y, x, parm, method = "weight", fagg = TRUE)
}
\arguments{
\item{y}{a vector (or a matrix) of outcomes.}

\item{x}{a vector (or a matrix) of covariates.}

\item{parm}{a vector (or a matrix) of the OR parameters.}

\item{method}{"weight" or "iterate" approach for baseline.}

\item{fagg}{a logical TRUE or FALSE for aggregate the y values
and corresponding baseline function values.}
}
\value{
\enumerate{
\item baseline function
2. corresponding y values
3. Location distinct y values in the ordered original y data
if aggregate option is set to TRUE
}
}
\description{
Two approaches are used to compute the baseline function estimator:
One uses the weighted (by the odds ratio function) approach
The other uses the iterative approach based on likelihood score
}
\details{
This function estimates the baseline function
given the OR parameters
}
\examples{
\dontrun{
# use the internal data file name dat, a 400x9 data matrix.
y=dat[,1:2], x=dat[,3:9];
fit=pwlkh(y,x)
parm=matrix(fit[[1]],nrow=dim(y)[2])
fit1=baseline(y,x,parm)
fit2=baseline(y,x,parm,fagg=TRUE)
fit3=baseline(y,x,parm,method='iterate')
}

}
\references{
Chen, H.Y. (2015). A note ogence of an iterative algorithm for semiparametric odds ratio models.
Biomatrika, 102, 747-751.

Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.
}
