% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprob.R
\name{cprob}
\alias{cprob}
\title{Compute conditional probability for a given covariate x, outcome y, OR parameter theta, and baseline function F.}
\usage{
cprob(y, x, parm, F)
}
\arguments{
\item{y}{a vector (or a matrix) of all possible outcome values.}

\item{x}{a vector of covariates.}

\item{parm}{a vector (or a matrix) of the OR parameters.}

\item{F}{baseline probabilities for all possible y values}
}
\value{
\enumerate{
\item all the predictive probabilities for a given covariate x.
}
}
\description{
Compute conditional probability for a given covariate x, outcome y, OR parameter theta, and baseline function F.
}
\details{
This function estimates the predictive probabilities for
given parameters theta and baseline function
}
\examples{
\dontrun{
# use the internal data file name dat, a 400x9 data matrix.
y=dat[,1:2], x=dat[,3:9];
fit=pwlkh(y,x)
theta=matrix(fit[[1]],nrow=dim(y)[2])
fit=baseline(y,x,parm,method='iterate')
fit1=cprob(y,x,parm=theta,F=fit[[1]])
predprob=fit1[[1]]
}

}
\references{
Chen, H.Y. (2015). A note ogence of an iterative algorithm for semiparametric odds ratio models.
Biomatrika, 102, 747-751.

Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.
}
