% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{This function is used for diagnostic of convergence}
\usage{
draw(x)
}
\arguments{
\item{x}{input argument.}
}
\description{
This function is used to draw the progression of the parameter estimates
in the iterations of the permutation likelihood approach to monitor the convergence
of the parameter estimates.
}
