% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmcem.R
\name{gmcem}
\alias{gmcem}
\title{Implementing Monte Carlo EM algorithm by coordinate-wise conditional maximization}
\usage{
gmcem(civ, method = "sp", nem = 10, nimpute = 1, stepsize = 0.5, nstep = 20)
}
\arguments{
\item{civ}{an array civ(n,p,2) with the same number of rows and columns as the
input data and the third dimension has length two to denote lower
and upper bounds of the censoring interval.
For example, if y_k is missing, then lower and upper bounds are -infty
and +infty; We use the observed min and max values to denote them respectively.
If y_k is left censored, the lower bound is set to min and the upper bound
is set to the observed bound. If y_k is right censored, the lower bound
is set to the observed bound and the right bound is set to max.
if y_k is interval censored, the lower bound and upper bound may be
respectively like 4 and 9, meaning the censored value is in the interval (4,9).}

\item{method}{specific method of estimation, can be
method='sp' for semiparametric likelihood approach (default)
method='pw' for pairwise likelihood approach
method='pm' for permutation likelihood approach}

\item{nem}{the number of EM steps}

\item{nimpute}{the number of Monte Carlo copies for each missing value}

\item{stepsize}{controls the step size of each M-step update.}

\item{nstep}{controls the number of steps before the change of step size}
}
\value{
\enumerate{
\item parameter estimate
2. one copy of the imputed data
3. history of parameter estimates
4. history of variance estimates
}
}
\description{
Implementing Monte Carlo EM algorithm by coordinate-wise conditional maximization
}
\details{
This function maximizes the conditional likelihood coordinate-wise
by the Monte Carlo EM algorithm. In M-step, all the OR parameters are
estimated by the semiparametric OR model using the semiparametric likelihood
approach based on the current imputed data. In MC-step, baseline is first
obtained. Imputing missing values are done coordinate-wise by
the sampling approach
}
\examples{
\dontrun{
 gmcem(civ=extbone, nem=100, stepsize=0.6,nstep=5)
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.
}
