% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertloc.R
\name{insertloc}
\alias{insertloc}
\title{The function finds the locations of setb elements in order of seta.
Both sequences are sorted from smallest to the largest}
\usage{
insertloc(seta, setb, method = "nearestleft")
}
\arguments{
\item{seta}{a set of distinctive elements ordered from smallest to the largest.}

\item{setb}{a set of distinctive elements ordered from smallest to the largest.}

\item{method}{either use "nearestleft" for low bound of interval censoring
or "nearestright" for upp bound of interval censoring}
}
\value{
a set of locations in sequence seta for the positions in setb
}
\description{
The function finds the locations of setb elements in order of seta.
Both sequences are sorted from smallest to the largest
}
\details{
This function identifies locations of setb in the sequence of seta
}
\examples{
\dontrun{
seta=c(-1,0,3,5,12,17,18)
setb=c(-9,-7,-5,2,2,3,4,5,8,15,17,17,23,27,29)
insertloc(seta,setb)
insertloc(seta,setb,method="nearestright")
}

}
