% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcem.R
\name{mcem}
\alias{mcem}
\title{Implementing Monte Carlo EM algorithm by coordinate-wise conditional maximization}
\usage{
mcem(
  dat = dat,
  miscode = c(-9),
  intvcen = FALSE,
  cintv = 0,
  method = "sp",
  nem = 10,
  nimpute = 1,
  stepsize = 0.3,
  nstep = 20
)
}
\arguments{
\item{dat}{a data matrix with possible missing values}

\item{miscode}{a set of values used for denoting missing values}

\item{intvcen}{The default intvcen==FALSE means only purely missing can occur. No need to
include cintv.
if intvcen=TRUE, the matrix cintv needs to be provided to denote the censoring interval
this can also be used for purely missing where cintv(,1)=1,cintv(,2)=n.}

\item{cintv}{an array cintv(n,p,2) with the same number of rows columns as the input data,
and the third dimension has length two to store the start and end of each
censoring interval. For example, if y_k is missing, then
the start and end numbers are (1,n); if y_k is censored, the start and end numbers
may be something like (4,9) meaning the censored value is in (y(4),y(9))}

\item{method}{specific method of estimation, can be
method='sp' for semiparametric likelihood approach (default)
method='pw' for pairwise likelihood approach
method='pm' for permutation likelihood approach}

\item{nem}{the number of EM steps}

\item{nimpute}{the number of Monte Carlo copies for each missing value}

\item{stepsize}{controls the step size of each M-step update.}

\item{nstep}{controls the number of steps before the change of step size}
}
\value{
\enumerate{
\item parameter estimate
2. one copy of the imputed data
}
}
\description{
Implementing Monte Carlo EM algorithm by coordinate-wise conditional maximization
}
\details{
This function maximizes the conditional likelihood coordinate-wise
by the Monte Carlo EM algorithm. In M-step, all the OR parameters are
estimated by the semiparametric OR model using the semiparametric likelihood
approach based on the current imputed data. In MC-step, baseline is first
obtained. Imputing missing values are done coordinate-wise by
the sampling approach
}
\examples{
\dontrun{
 mcem(dat=bone,miss=c(-9),nimpute=100)
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.
}
