% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdimpute.R
\name{mdimpute}
\alias{mdimpute}
\title{Impute missing values by random draw from the marginal distributions}
\usage{
mdimpute(dat = dat, miscode = c(-9), nimpute = 5)
}
\arguments{
\item{dat}{a data matrix with possible missing values}

\item{miscode}{a set of values used for denoting missing values}

\item{nimpute}{the number of copies of the imputed data}
}
\value{
\enumerate{
\item Imputed data set with the last variable indexing the datasets
}
}
\description{
This is a simple impute to create initial values for missing data
}
\details{
This function imputes missing data by random draw from
the marginal distribution of the respective variables
}
\examples{
\dontrun{
 fit=mdimpute(dat=bone,miss=c(-9))
 fit[[1]]
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.
}
