% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkselection.R
\name{networkselect}
\alias{networkselect}
\title{Network selection}
\usage{
networkselect(dat, group, lambda, network, method, criterion)
}
\arguments{
\item{dat}{the whole data matrix.}

\item{group}{a vector indicating the sizes of the variables in each group the data matrix divides.}

\item{lambda}{a vector of penalty parameters. A network is selected for each lambda value.}

\item{network}{a set of networks correspond to lambda vector.}

\item{method}{the method used to fit the network model, choices include: "pw" for pairwise pseudo-likelihood
approach, "pm" for permutation likelihood approach, and "sp" for semiparametric likelihood approach.}

\item{criterion}{the network selection criterion: "BIC".}
}
\value{
The function returns

1. the index of the selected model

2. the structure of the selected network

3. all the BICs corresponding to the set of networks

4. all the log-likelihood values corresponding to the set of networks: sum of the node-wise pseudo-likelihoods for
\code{method = "pw"}; sum of the node-wise semiparametric likelihoods for \code{method = "sp"}; the joint permutation
likelihoods for \code{method = "pm"}
}
\description{
This function selects one network among a set of networks using a specified criterion based on a particular
fitting approach.
}
\details{
One has to specify the whole data matrix including the outcome and covariates, a vector indicating the sizes of
variables in each group that data matrix divides has to be specified.

One can choose the desired approach, which including "pw", "pm", or "sp" for fitting the network model. Alsom a network
selection criterion has to be specified. Currently, the implemented network selection criterion is "BIC".
}
\examples{
\dontrun{
n <- 200; p <- 10
datmat <- matrix(rnorm(n * p), ncol = p)
vargroup <- c(2, 3, 1, 4)
penaltyparam <- c(10, 50, 500)
fit <- pmpenlkh(dat = datmat, group = vargroup, lambda = penaltyparam,
                nsamp = 2e4, niter = 50)
network <- fit[[4]]
fit0 <- networkselect(dat = datmat, group = vargroup, lambda = penaltyparam,
                      network = network, method = 'pm', criterion = 'BIC')
fit0[[2]]
}

}
\references{
Chen, H. Y. and Chen, J. (2021). Semiparametric odds ratio model for network detection. Manuscript.

Chen, H. Y. (2021). Semiparametric odds ratio model and its applications. draft.
}
