% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmlkh.R
\name{pmlkh}
\alias{pmlkh}
\title{The simple permutation likelihood approach}
\usage{
pmlkh(
  dat,
  group,
  niter = 50,
  eps = 0.01,
  nlag = 20,
  nburnin = 50000,
  nsamp = 1e+05,
  nintv = 50,
  maxcyc = 2
)
}
\arguments{
\item{dat}{a data matrix of nxp dimension,}

\item{group}{a vector of positive integers of length ng
such that sum(group)=p.}

\item{niter}{the maximum number of iterations}

\item{eps}{convergence criterion: discrepancy between successive
iterations for terminal the iteration.}

\item{nlag}{the averaged parameter values in the sequence to assess convergence, default nlag=20.}

\item{nburnin}{the burnin steps before a sample is taken. default nburnin=5e4.}

\item{nsamp}{the sample sizes to be taken in the MCMC sampling of permutation. default nsamp=1e4.}

\item{nintv}{the sampling interval (number of steps ignored) in the MCMC chain. default nintv=2e2.}

\item{maxcyc}{the maximum step size (cycle of the permutation) in obtaining a candidate permutation
(change from the previous one) in the Gibbs sampling. default maxcyc=2.}
}
\value{
Estimate of the model parameters (column-wise vectorization) and
        the covariance matrix estimate corresponding to the vectorized parameters.
}
\description{
This approach uses the maximum permutation likelihood approach to estimating
the parameters in the semiparametric odds ratio model.
}
\details{
This method maximizes the permutation likelihood to obtain the parameter estimator
         and uses the inverse of the permutation likelihood information matrix to estimate
         the asymptotic variance of the estimator.
         The permutation likelihood is approximated using the Gibbs sampler
         of permutations with non-constant weights (probability).
}
\examples{
\dontrun{
n=200; p=10
datmat=matrix(rnorm(n * p), ncol = p)
vargroup=c(2, 3, 1, 4)
pmlkh(dat=datmat, group=vargroup)
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.
}
