% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwhdlkh.R
\name{pwhdlkh}
\alias{pwhdlkh}
\title{The penalized pairwise pseudo-likelihood approach for model selection}
\usage{
pwhdlkh(y, x, lambda, niter = 50, eps = 1e-06)
}
\arguments{
\item{y}{a vector of outcomes}

\item{x}{a vector of covariates}

\item{lambda}{a vector of penalty values for model selection,
each penalty value determine a model.}

\item{niter}{maximum number of iterations in finding the estimator.}

\item{eps}{convergence criterion: discrepancy between successive
iterations for terminal the iteration}
}
\value{
a set of selected parameters to determine the model.
}
\description{
This approach selects a model by parameter selection
under the semiparametric odds ratio model
using the penalized pairwise pseudo-likelihood approach.
}
\details{
This method maximizes the penalized pairwise pseudo-likelihood
with penalty being the l^1-norm of the parameters to select
a sparse model.
}
\examples{
\dontrun{
# use the internal data file name dat, a 400x9 data matrix.
y=dat[,1:2], x=dat[,3:9];lambda=c(10,50,200,500,800)
pwHDlkh(y,x,lambda=lambda)
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.

Liang, K.Y. and Qin, J. (2000). Regression analysis under non-standard situations:
a pairwise pseudolikelihood approach, Journal of the Royal Statistical Society, Ser. B,
62, 773-786.
}
