% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwLKH.R
\name{pwlkh}
\alias{pwlkh}
\alias{pwlkhfix}
\alias{pwpenlkh}
\title{Pairwise pseudo-likelihood approach}
\usage{
pwlkh(y, x, niter = 50, eps = 1e-06, vect = "col")

pwlkhfix(y, x, fixstruct, niter = 50, eps = 1e-06, vect = "col")

pwpenlkh(dat, group, lambda, niter = 50, eps = 1e-06)
}
\arguments{
\item{y}{the outcome data matrix. This must be specified.}

\item{x}{the covariate data matrix. This must be specified.}

\item{niter}{the maximum number of iterations.}

\item{eps}{convergence criterion.}

\item{vect}{the method of vectoring parameter matrix. Either column-wise or row-wise.}

\item{dat}{the whole data matrix.}

\item{lambda}{a vector of penalty parameters. A network is selected for each lambda value.}

\item{group}{a vector denoting the sizes of the variables in each group the data matrix divided.}

\item{fixstruct}{a matrix indicating which coefficients are set to 0.}
}
\value{
The \code{pwlkh} function returns

1. the odds ratio parameter estimates.

2. the estimated variance-covariance matrix of the odds ratio estimator.

The \code{pwlkhfix} function returns

1. the odds ratio parameter estimates.

2. the estimated variance-covariance matrix of the odds ratio estimator.

3. the pairwise pseudo-likelihood value.

The \code{pwpenlkh} function returns

1. The group vector specified.

2. A set of networks selected by the penalized pseudo-likelihood with within group labeled by NA.

3. Annotation of the entries of the network structure matrix.

4. A set of networks selected by the penalized pseudo-likelihood with within group labeled by 0.
}
\description{
The \code{pwlkh} and \code{pwlkhfix} functions estimate the odds ratio parameters in the semiparametric odds ratio
model using the pairwise pseudo-likelihood approach proposed in Liang and Qin (2000).
The \code{pwpenlkh} function performs the node-wise selection of the network based on the semiparametric odds ratio
model. Currently, only the bilinear log-odds ratio function is allowed.
}
\details{
The \code{pwlkhfix} function is more general than the \code{pwlkh} function since the former not only allows for
the fixed zero coefficients, it also computes the logarithm of the pairwise pseudo-likelihood values.

While the outcome data matrix and the variate data matrix has to be specified separately in the \code{pwlkh} and
the \code{pwlkhfix} functions, for the \code{pwpenlkh} function, one has to specify the whole data matrix including
the outcome and the covariates.

For the \code{pwlkhfix} function, one has to specify a matrix for a fixed structure, that indicating which
coefficients are set to zero. For the \code{pwpenlkh} function, a \code{lambda} parameter that specifies the
penalties on the pairwise pseudo-likelihood, is required.
}
\examples{

n <- 200; p <- 3; q <- 4
covariates <- matrix(rnorm(n * q), ncol = q)
outcomes <- covariates \%*\% matrix(runif(q * p), ncol = p) + matrix(rnorm(n * p), ncol = p)

# pwlkh
pwlkh(y = outcomes, x = covariates)

# pwlkhfix
structure <- matrix(rbind(p * q), ncol = q)
pwlkhfix(y = outcomes, x = covariates, fixstruct = structure)

# pwpenlkh
n <- 200; p <- 10
datmat <- matrix(rnorm(n * p), ncol = p)
vargroup <- c(2, 3, 1, 4)
penaltyparam <- c(10, 50, 500)
pwpenlkh(dat = datmat, group = vargroup, lambda = penaltyparam)

}
\references{
Liang, K. Y. and Qin, J. (2000). Regression analysis under non-standard situations: a
  pseudolikelihood approach. Journal of the Royal Statistical Society, Ser. B, 62:773-786.

Chen, H. Y. (2007). A semiparametric odds ratio model for measuring association.
   Biometrics, 63: 413-421.

...
}
