% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwmodelselect.R
\name{pwmodelselect}
\alias{pwmodelselect}
\title{Penalized model selection by L^1 penalized pairwise pseudo-likelihood
with AIC type of criterion for tuning parameter selection}
\usage{
pwmodelselect(y, w, lamb)
}
\arguments{
\item{y}{a vector of outcomes}

\item{w}{a vector of covariates}

\item{lamb}{a vector of penalty values for model selection,
each penalty value determine a model.}
}
\value{
a set of selected parameters to determine the model, AIC values,
and the structure of the model
}
\description{
This approach selects an odds ratio model
using the penalized pairwise pseudo-likelihood approach
with AIC for tuning parameter selection.
}
\details{
The final selected model is determined by refitting
using AIC criterion.
}
\examples{
\dontrun{
# use the internal data file name dat, a 400x9 data matrix.
y=dat[,1:2], x=dat[,3:9];lambda=c(10,50,200,500,800)
pwmodelselect(y,w,lamb)
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.

Liang, K.Y. and Qin, J. (2000). Regression analysis under non-standard situations:
a pairwise pseudolikelihood approach, Journal of the Royal Statistical Society, Ser. B,
62, 773-786.
}
