% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwpenlkh.R
\name{pwpenlkh}
\alias{pwpenlkh}
\title{The penalized pairwise pseudo-likelihood approach for node-wise network selection}
\usage{
pwpenlkh(dat, group, lambda, niter = 50, eps = 1e-06)
}
\arguments{
\item{dat}{a data matrix of nxnode dimension,}

\item{group}{a vector of positive integers of length ng
such that sum(group)=node.}

\item{lambda}{a vector of penalty values for network selection,
each penalty value determine a network.}

\item{niter}{maximum number of iterations in finding the estimator.}

\item{eps}{convergence criterion: discrepancy between successive
iterations for terminal the iteration}
}
\value{
group denoting the input node clusters, a set of networks identified corresponding
          to the penalty values (with within group connections coded by NA),
          annotation of network connection (0=not connected etc.)
          the set of networks identified with within group connections coded by 0.
}
\description{
This approach selects network structure
   under the semiparametric odds ratio model
   using the penalized pairwise pseudo-likelihood approach.
   The network selection is a node-wise approach.
}
\details{
This method maximizes the penalized pairwise pseudo-likelihood
         with penalty being the l^1-norm of the parameters to select
         the network structure.
}
\examples{
\dontrun{
# use the internal data file name dat, a 400x9 data matrix.
group=c(2,3,4);lambda=c(10,50,200,500,800)
pwpenlkh(dat=dat,group=group,lambda=lambda)
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.

Liang, K.Y. and Qin, J. (2000). Regression analysis under non-standard situations:
            a pairwise pseudolikelihood approach, Journal of the Royal Statistical Society, Ser. B,
            62, 773-786.
}
