% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spLKH.R
\name{splkh}
\alias{splkh}
\alias{splkhfix}
\alias{sppenlkh}
\title{Semiparametric likelihood approach}
\usage{
splkh(y, x, niter = 50, eps = 1e-06, vect = "col")

splkhfix(y, x, fixstruct, niter = 50, eps = 1e-06, vect = "col")

sppenlkh(dat, group, lambda, niter = 50, eps = 1e-06)
}
\arguments{
\item{y}{the outcome data matrix. This must be specified.}

\item{x}{the covariate data matrix. This must be specified.}

\item{niter}{the maximum number of iterations.}

\item{eps}{convergence criterion.}

\item{vect}{the method of vectoring parameter matrix. Either column-wise or row-wise.}

\item{fixstruct}{a matrix indicating which coefficients are set to 0.}

\item{dat}{the whole data matrix.}

\item{group}{a vector denoting the sizes of the variables in each group the data matrix divided.}

\item{lambda}{a vector of penalty parameters. A network is selected for each lambda value.}
}
\value{
The \code{splkh} function returns

1. the odds ratio parameter estimates.

2. the estimated variance-covariance matrix of the odds ratio estimator.

The \code{splkhfix} function returns

1. the odds ratio parameter estimates.

2. the estimated variance-covariance matrix of the odds ratio estimator.

3. logarithm of the maximum semiparametric likelihood.

The \code{sppenlkh} returns

1. The group vector specified.

2. A set of networks selected by the penalized pseudo-likelihood with within group labeled by NA.

3. Annotation of the entries of the network structure matrix.

4. A set of networks selected by the penalized pseudo-likelihood with within group labeled by 0.
}
\description{
The \code{splkh} and the \code{splkhfix} functions estimate the odds ratio parameters in the semiparametric odds
ratio model using the maximum semiparametric likelihood approach proposed in Chen et al. (2015). The \code{sppenlkh}
function performs the network selection using the penalized sepiparametric likelihood approach based on the
semiparametric odds ratio model. Currently, only the bilinear log-odds ratio function is allowed.
}
\details{
The \code{splkhfix} function is more general than the \code{splkh} function since the former not only allows for
the fixed zero coefficients, it also computes the logarithm of the maximum semiparametric likelihood values.

While the outcome data matrix and the variate data matrix have to be specified separately in the \code{splkh} and
the \code{splkhfix} functions, for the \code{sppenlkh} function, one has to specify the whole data matrix including
the outcome and the covariates.

For the \code{splkhfix} function, one has to specify a matrix for a fixed structure, that indicating which
coefficients are set to zero. For the \code{sppenlkh} function, a \code{lambda} parameter that specifies the
penalties on the semiparametric likelihood, is required.
}
\examples{

n <- 200; p <- 3; q <- 4
covariates <- matrix(rnorm(n * q), ncol = q)
outcomes <- covariates \%*\% matrix(runif(q * p), ncol = p) + matrix(rnorm(n * p), ncol = p)

# splkh
splkh(y = outcomes, x = covariates)

# splkhfix
structure <- matrix(rbind(p * q), ncol = q)
splkhfix(y = outcomes, x = covariates, fixstruct = structure)

# sppenlkh
n <- 200; p <- 10
datmat <- matrix(rnorm(n * p), ncol = p)
vargroup <- c(2, 3, 1, 4)
penaltyparam <- c(10, 50, 500)
sppenlkh(dat = datmat, group = vargroup, lambda = penaltyparam)

}
\references{
...

...
}
