% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sporm-package.R
\docType{package}
\name{Sporm-package}
\alias{Sporm}
\alias{Sporm-package}
\title{Sporm: Statistical analysis under the Semi-Parametric Odds Ratio Model}
\description{
This package implements multivariate analysis and variable selection for 
    multivariate data based on various likelihood approach.
}
\keyword{internal}
