% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bone.R
\docType{data}
\name{bone}
\alias{bone}
\title{Bone fracture data set}
\format{
A data matrix with 436 records and 28 variables. -9 in any variable
denotes missing value. race and age are matching variables.
}
\source{
A case-control study of hip fractures in veteran,
}
\usage{
bone
}
\description{
This data set contains data from the case-control study of hip fracture
in veterans.
}
\keyword{datasets}
