% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{A 400x9 matrix to test the function within the package}
\format{
A data matrix whose dimension is 400x9.
}
\usage{
dat
}
\description{
A 400x9 matrix to test the function within the package.
}
\keyword{datasets}
