% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{Example data set}
\format{
A data matrix with 400 records and 9 variables.
}
\source{
simulated data
}
\usage{
example
}
\description{
Example data is a 400x9 matrix to test the function within the package.
}
\keyword{datasets}
