% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcem.R
\name{mcem}
\alias{mcem}
\title{Implementing Monte Carlo EM algorithm by coordinate-wise conditional maximization}
\usage{
mcem(
  dat = dat,
  miscode = c(-9),
  method = "sp",
  nem = 10,
  nimpute = 5,
  nseq = 10
)
}
\arguments{
\item{dat}{a data matrix with possible missing values}

\item{miscode}{a set of values used for denoting missing values}

\item{method}{specific method of estimation, can be
method='sp' for semiparametric likelihood approach (default)
method='pw' for pairwise likelihood approach
method='pm' for permutation likelihood approach}

\item{nem}{the number of EM steps}

\item{nimpute}{the number of Monte Carlo copies for each missing value}

\item{nseq}{the number of repeated sampling steps in each MC-step}
}
\value{
\enumerate{
\item parameter estimate
2. one copy of the imputed data
}
}
\description{
Implementing Monte Carlo EM algorithm by coordinate-wise conditional maximization
}
\details{
This function maximizes the conditional likelihood coordinate-wise
by the Monte Carlo EM algorithm. In M-step, all the OR parameters are
estimated by the semiparametric OR model using the semiparametric likelihood
approach based on the current imputed data. In MC-step, baseline is first
obtained. Imputing missing values are done coordinate-wise by
the sampling approach
}
\examples{
\dontrun{
 mcem(dat=bone,miss=c(-9),nimpute=100)
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.
}
