% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkselection.R
\name{networkselect}
\alias{networkselect}
\title{The network selection from a set of selected networks using output from one of
the penalized likelihood approaches}
\usage{
networkselect(dat, group, lambda, network, method, criterion)
}
\arguments{
\item{dat}{a data matrix of nxnode dimension}

\item{group}{a vector of positive integers of length ng such that sum(group) = node}

\item{lambda}{a vector of penalty values for network selection, each penalty value determines a network}

\item{network}{a set of networks determined by one of the penalized approaches}

\item{method}{the method used in the likelihood approach ("pw", "sp", or "pm")}

\item{criterion}{selection criterion such as BIC}
}
\value{
Estimate of the model parameters (vectorized) and
        the covariance matrix estimate corresponding to the vectorized parameters.
}
\description{
This approach selects the network using the BIC.
}
\details{
This approach selects the network using the BIC, by refitting each
network to obtain the log-likelihood (objective function) and then
uses BIC for the selection.
}
\examples{
\dontrun{
n <- 200; p <- 10
datmat <- matrix(rnorm(n * p), ncol = p)
vargroup <- c(2, 3, 1, 4)
penaltyparam <- c(10, 50, 500)
fit <- pmpenlkh(dat = datmat, group = vargroup, lambda = penaltyparam,
                nsamp = 2e4, niter = 50)
network <- fit[[4]]
fit0 <- networkselect(dat = datmat, group = vargroup, lambda = penaltyparam,
                      network = network, method = 'pm', criterion = 'BIC')
fit0[[2]]
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.

Chen, H. Y. and Chen, J. (2020). Network selection through semiparametric odds ratio model. Manuscript.
}
