% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmlkh.R
\name{pmlkh}
\alias{pmlkh}
\title{The simple permutation likelihood approach}
\usage{
pmlkh(
  dat,
  group,
  niter = 50,
  eps = 0.01,
  nlag = 20,
  plot = TRUE,
  nburnin = 50000,
  nsamp = 1e+05,
  nintv = 50,
  maxcyc = 2,
  inparm = FALSE,
  theta = 0,
  steplen = c(1, 1)
)
}
\arguments{
\item{dat}{a data matrix of nxp dimension.}

\item{group}{a vector of positive integers of length ng
such that sum(group)=p.}

\item{niter}{the maximum number of iterations.}

\item{eps}{convergence criterion: discrepancy between successive
iterations for terminating the iteration.}

\item{nlag}{the averaged parameter values in the sequence to assess convergence, default nlag=20.}

\item{plot}{Default is TRUE, to produce the convergence plot. If FALSE, suppress the plot.}

\item{nburnin}{the burnin steps before a sample is taken, default nburnin=5e4.}

\item{nsamp}{the sample sizes to be taken in the MCMC sampling of permutation, default nsamp=1e4.}

\item{nintv}{the sampling interval (number of steps ignored) in the MCMC chain, default nintv=2e2.}

\item{maxcyc}{the maximum step size (cycle of the permutation) in obtaining a candidate permutation
(change from the previous one) in the Gibbs sampling, default maxcyc=2.}

\item{inparm}{indicate whether theta is initialized in the input}

\item{theta}{provide theta value if theta is initialized.}

\item{steplen}{two-stage controls step length of the parameter update,steplen=1 means no control, steplen<1 means control.
This parameter is set to make convergence more stable. the first component of steplen controls the step size in 1 to 15 iterations,}
}
\value{
Estimate of the model parameters (column-wise vectorization) and
the covariance matrix estimate corresponding to the vectorized parameters.
}
\description{
This approach uses the maximum permutation likelihood approach to estimating
the parameters in the semiparametric odds ratio model.
}
\details{
This method maximizes the permutation likelihood to obtain the parameter estimator
and uses the inverse of the permutation likelihood information matrix to estimate
the asymptotic variance of the estimator.
The permutation likelihood is approximated using the Gibbs sampler
of permutations with non-constant weights (probability).
}
\examples{
\dontrun{
pmlkh(dat=example, group=c(1,8))
pmlkh(dat=example, group=rep(1,9),nburn=1000,niter=100,steplen=c(0.2,1))
}

}
\references{
Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.
}
