% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwlkh.R
\name{pwlkh}
\alias{pwlkh}
\title{The simple pairwise pseudo-likelihood approach}
\usage{
pwlkh(
  y,
  x,
  niter = 50,
  eps = 1e-06,
  vect = "col",
  steplen = c(0.5, 1),
  nstep = 10
)
}
\arguments{
\item{y}{outcomes: can be a vector of length n or a matrix of nxq dimension.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{niter}{maximum number of iterations in finding the estimator.}

\item{eps}{convergence criterion: discrepancy between successive
iterations for terminating the iteration.}

\item{vect}{methods of vectorization of parameter matrix.
Either column-wise (vect='col') or row-wise (vect='row').}

\item{steplen}{two-stage controls step length of the parameter update,steplen=1 means no control,
steplen<1 means control.This parameter is set to make convergence more stable. the first
component of steplen controls the step size in 1 to nstep iterations, the 2nd component
of steplen controls the stepsize for the remaining iterations.}

\item{nstep}{controls the number of steps before a change of the step size occurs}
}
\value{
Estimate of the model parameters (vectorized) and
the covariance matrix estimate corresponding to the vectorized parameters.
}
\description{
This approach uses the maximum pairwise pseudo-likelihood approach to estimating
the parameters in the semiparametric odds ratio model.
}
\details{
This method maximizes the pairwise likelihood to obtain the parameter estimator
and uses U-statistic theory to estimate the asymptotic variance of the estimator.
}
\examples{
\dontrun{
Example 1:
n=100; p=2
x=matrix(2*runif(n*p)-1,ncol=p)
y=x[,1]-x[,2]+rnorm(n)
pwlkh(y,x,niter=50,eps=1e-6,vect='col')

Example 2:
n=200;p=2;q=2
x=matrix(2*runif(n*p)-1,ncol=p)
y=array(0,c(n,q))
y[,1]=x[,1]-x[,2]+rnorm(n)
y[,2]=x[,1]+x[,2]+rnorm(n)
pwlkh(y,x,niter=40,eps=1e-6,vect="row")
}

}
\references{
Chan K. C. G. (2013). Nuisance parameter elimination for proportional likelihood
ratio models with nonignorable missingness and random truncation. \emph{Biometrika}, \strong{100},
269-276.

Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.

Liang, K.Y. and Qin, J. (2000). Regression analysis under non-standard situations:
a pairwise pseudolikelihood approach. \emph{Journal of the Royal Statistical Society, Ser. B},
\strong{62}, 773-786.
}
