% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splkh.R
\name{splkh}
\alias{splkh}
\title{The simple semiparametric likelihood approach}
\usage{
splkh(y, x, niter = 50, eps = 1e-06, vect = "col")
}
\arguments{
\item{y}{outcomes: can be a vector of length n or a matrix of nxq dimension,}

\item{x}{covariates: a matrix of nxp dimension.}

\item{niter}{maximum number of iterations in finding the estimator.}

\item{eps}{convergence criterion: discrepancy between successive
iterations for terminal the iteration}

\item{vect}{methods of vectorization of parameter matrix.
Either column-wise (vect='col') or row-wise (vect='row').}
}
\value{
Estimate of the model parameters (vectorized) and
        the covariance matrix estimate corresponding to the vectorized parameters.
}
\description{
This approach uses the maximum semiparametric likelihood approach to estimating
the parameters in the semiparametric odds ratio model.
}
\details{
This method maximizes the semiparametric likelihood to obtain the parameter estimator
         and uses the inverse of the profile information matrix to estimate
         the asymptotic variance of the estimator.
}
\examples{
\dontrun{Example 1:
n=100; p=2
x=matrix(2*runif(n*p)-1,ncol=p)
y=x[,1]-x[,2]+rnorm(n)
splkh(y,x,niter=50,eps=1e-6,vect='col')

Example 2:
n=200;p=2;q=2
x=matrix(2*runif(n*p)-1,ncol=p)
y=array(0,c(n,q))
y[,1]=x[,1]-x[,2]+rnorm(n)
y[,2]=x[,1]+x[,2]+rnorm(n)
splkh(y,x,niter=40,eps=1e-6,vect="row")}

}
\references{
Chen, H. Y. (2007). A semiparametric odds ratio model for measuring association,
            Biometrics,63,413--421.

Chen, H.Y. (2022). Semiparametric Odds Ratio Model and its Application. CRC press.

Chen, H. Y., and Rader, D. E., and Li, M. (2015). Likelihood inference on
             semiparametric odds ratio Model. Journal of the American Statistical Association,
             110,1125-1135.
}
