% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPORM-package.R
\docType{package}
\name{SPORM-package}
\alias{SPORM}
\alias{SPORM-package}
\title{SPORM: Statistical analysis under the Semi-Parametric Odds Ratio Model}
\description{
This package implements multivariate analysis and variable selection for 
    multivariate data based on various likelihood approach.
}
\keyword{internal}
